Delete Items Permissions is fully functional

+++ backend/src/resolvers/Mutation.js
@@ -47,8 +47,16 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // Find the item
-    const item = await ctx.db.query.item({ where }, `{ id title }`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
     // TODO: check if they own that item, or have the permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions
+      .some(permission => ['ADMIN', 'ITEMDELETE']
+        .includes(permission));
+
+    if (!ownsItem && hasPermissions) {
+      throw new Error('You don\'t have permission!')
+    };
 
     // Delete it
     return ctx.db.mutation.deleteItem({ where }, info);
